Persona terminada

+++ src/main/java/com/crowdfunding/ecofin/services/PersonaServices.java
@@ -25,16 +25,16 @@ public class PersonaServices {
 
     @Value("${jwt.secret}")
     private  String secretKey;
-
+    //5 * 60 * 1000 (5 minutos)
     private static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60 * 1000;
 
     private String getJWTToken(String id, String username, List<String> roles) {
         List<GrantedAuthority> grantedAuthorities = new ArrayList<>();
+
         for(String rol :roles){
             grantedAuthorities.add(new SimpleGrantedAuthority(rol));
         }
-        //List grantedAuthorities = AuthorityUtils.commaSeparatedStringToAuthorityList(roles.toString());
-        System.out.println(grantedAuthorities);
+
         String token = Jwts.builder().setId(id).setSubject(username)
                 .claim("authorities", grantedAuthorities)
                 .setIssuedAt(new Date(System.currentTimeMillis()))
@@ -47,14 +47,23 @@ public class PersonaServices {
     public PersonaDTO savePersona(PersonaDTO persona){ return  personaRepository.insert(persona); }
 
     public PersonaDTO findById(String id){
-        return personaRepository.findById(id);
+        if(id != null && !id.isEmpty())
+            return personaRepository.findById(id);
+        return null;
     }
 
-    public Map<String, Object>  login(String email, String password){
+    public Map<String, Object> login(String email, String password){
+
+        if(email == null || email.isEmpty() || password == null || password.isEmpty()){
+            return null;
+        }
+
         PersonaDTO persona = personaRepository.findByEmailAndPassword(email, password);
-        Map<String, Object> res = new HashMap<>();
+        Map<String, Object> res = null;
         if(persona != null){
+            res = new HashMap<>();
             res.put("TOKEN", getJWTToken(persona.getId(), persona.getNombre(), persona.getPerfil()));
+            res.put("TOKEN_ID", persona.getId());
             res.put("NOMBRE", persona.getNombre());
             res.put("EMAIL",persona.getEmail());
         }
@@ -65,25 +74,54 @@ public class PersonaServices {
         if(!persona.getId().isEmpty()){
             PersonaDTO actual = personaRepository.findById(persona.getId());
 
-            actual.setNombre(persona.getNombre());
-            actual.setEmail(persona.getEmail());
-            actual.setPerfil(persona.getPerfil());
-            actual.setPassword(persona.getPassword());
-            actual.setDireccion(persona.getDireccion());
-            actual.setProfesion(persona.getProfesion());
-            actual.setFechaNacimiento(persona.getFechaNacimiento());
-            actual.setSexo(persona.getSexo());
-            actual.setPais(persona.getPais());
-            actual.setActivo(persona.getActivo());
-            actual.setProyectos(persona.getProyectos());
+            if(persona.getNombre() != null){
+                actual.setNombre(persona.getNombre());
+            }
 
-            return personaRepository.save(persona);
+            if(persona.getEmail() != null){
+                actual.setEmail(persona.getEmail());
+            }
+
+            if(persona.getPerfil() != null){
+                actual.setPerfil(persona.getPerfil());
+            }
+
+            if(persona.getPassword() != null){
+                actual.setPassword(persona.getPassword());
+            }
+
+            if(persona.getDireccion() != null){
+                actual.setDireccion(persona.getDireccion());
+            }
+
+            if(persona.getProfesion() != null){
+                actual.setProfesion(persona.getProfesion());
+            }
+
+            if(persona.getFechaNacimiento() != null){
+                actual.setFechaNacimiento(persona.getFechaNacimiento());
+            }
+
+            if(persona.getSexo() != null){
+                actual.setSexo(persona.getSexo());
+            }
+
+            if(persona.getPais() != null){
+                actual.setPais(persona.getPais());
+            }
+
+            if(persona.getProyectos() != null){
+                actual.setProyectos(persona.getProyectos());
+            }
+
+            return personaRepository.save(actual);
         }
         return null;
     }
 
     public void deletePersona(String id){
-        personaRepository.deleteById(id);
+        if(id != null && !id.isEmpty())
+            personaRepository.deleteById(id);
     }
 
     public List<PersonaDTO> consultAllPersonas(){
